fix mobile version, handle deconection if token expired

+++ src/helpers/getDataFromTimestamp.js
@@ -1,8 +1,8 @@
 export const getDataFromTimestamp = (timestamp) => {
   const options = {
-    weekday: 'long',
+    weekday: 'short',
     year: 'numeric',
-    month: 'long',
+    month: 'short',
     day: 'numeric',
     hour: '2-digit',
     minute: '2-digit',
